Update resources/js and app folders from tutorial

+++ app/Http/Controllers/FileController.php
@@ -5,37 +5,34 @@ namespace App\Http\Controllers;
 use App\Http\Requests\FilesActionRequest;
 use App\Http\Requests\StoreFileRequest;
 use App\Http\Requests\StoreFolderRequest;
-use App\Http\Requests\UpdateFileRequest;
 use App\Http\Resources\FileResource;
 use App\Models\File;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 use Inertia\Inertia;
-use ZipArchive;
 
 class FileController extends Controller
 {
-
     public function myFiles(Request $request, string $folder = null)
     {
         if ($folder) {
-            $folder = File::where('created_by', '=', request()->user()->id)
-                ->where('path', '=', $folder)
+            $folder = File::query()->where('created_by', Auth::id())
+                ->where('path', $folder)
                 ->firstOrFail();
         }
-
         if (!$folder) {
             $folder = $this->getRoot();
         }
-        $files = FileResource::collection(
-            File::query()
-                ->where('parent_id', '=', $folder->id)
-                ->where('created_by', '=', request()->user()->id)
-                ->orderBy('is_folder', 'desc')
-                ->orderBy('created_at', 'desc')
-                ->paginate(10)
-        );
+        $files = File::query()
+            ->where('parent_id', $folder->id)
+            ->where('created_by', Auth::id())
+            ->orderBy('is_folder', 'desc')
+            ->orderBy('created_at', 'desc')
+            ->paginate(10);
+
+        $files = FileResource::collection($files);
 
         if ($request->wantsJson()) {
             return $files;
@@ -43,48 +40,27 @@ class FileController extends Controller
 
         $ancestors = FileResource::collection([...$folder->ancestors, $folder]);
 
-        if ($folder) {
-            $folder = new FileResource($folder);
-        }
+        $folder = new FileResource($folder);
 
-        return Inertia::render('File/MyFiles', compact('files', 'folder', 'ancestors'));
-    }
-
-    public function sharedWithMe()
-    {
-        return Inertia::render('File/MyFiles');
-    }
-
-    public function sharedByMe()
-    {
-        return Inertia::render('File/MyFiles');
-    }
-
-    public function trash()
-    {
-        return Inertia::render('File/MyFiles');
+        return Inertia::render('MyFiles', compact('files', 'folder', 'ancestors'));
     }
 
     public function createFolder(StoreFolderRequest $request)
     {
         $data = $request->validated();
-
         $parent = $request->parent;
 
         if (!$parent) {
             $parent = $this->getRoot();
         }
 
-        $folder = new File();
-        $folder->is_folder = true;
-        $folder->name = $data['name'];
+        $file = new File();
+        $file->is_folder = 1;
+        $file->name = $data['name'];
 
-        $parent->appendNode($folder);
+        $parent->appendNode($file);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(StoreFileRequest $request)
     {
         $data = $request->validated();
@@ -100,32 +76,35 @@ class FileController extends Controller
             $this->saveFileTree($fileTree, $parent, $user);
         } else {
             foreach ($data['files'] as $file) {
-                /** @var $file \Illuminate\Http\UploadedFile */
+                /** @var \Illuminate\Http\UploadedFile $file */
 
-                $path = $file->store('/files/' . $user->id);
-                $model = new File();
-                $model->storage_path = $path;
-                $model->is_folder = false;
-                $model->name = $file->getClientOriginalName();
-                $model->mime = $file->getMimeType();
-                $model->size = $file->getSize();
-
-                $parent->appendNode($model);
+                $this->saveFile($file, $user, $parent);
             }
         }
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(UpdateFileRequest $request, File $file)
+    private function getRoot()
     {
-        //
+        return File::query()->whereIsRoot()->where('created_by', Auth::id())->firstOrFail();
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
+    public function saveFileTree($fileTree, $parent, $user)
+    {
+        foreach ($fileTree as $name => $file) {
+            if (is_array($file)) {
+                $folder = new File();
+                $folder->is_folder = 1;
+                $folder->name = $name;
+
+                $parent->appendNode($folder);
+                $this->saveFileTree($file, $folder, $user);
+            } else {
+
+                $this->saveFile($file, $user, $parent);
+            }
+        }
+    }
+
     public function destroy(FilesActionRequest $request)
     {
         $data = $request->validated();
@@ -137,10 +116,13 @@ class FileController extends Controller
             foreach ($children as $child) {
                 $child->delete();
             }
-        }
-        foreach (( $data['ids'] ?? [] ) as $id) {
-            $file = File::find($id);
-            $file->delete();
+        } else {
+            foreach ($data['ids'] ?? [] as $id) {
+                $file = File::find($id);
+                if ($file) {
+                    $file->delete();
+                }
+            }
         }
 
         return to_route('myFiles', ['folder' => $parent->path]);
@@ -153,14 +135,13 @@ class FileController extends Controller
 
         $all = $data['all'] ?? false;
         $ids = $data['ids'] ?? [];
-        if (!$all & empty($ids)) {
+
+        if (!$all && empty($ids)) {
             return [
                 'message' => 'Please select files to download'
             ];
         }
 
-        $url = null;
-        $filename = null;
         if ($all) {
             $url = $this->createZip($parent->children);
             $filename = $parent->name . '.zip';
@@ -176,11 +157,10 @@ class FileController extends Controller
                     $url = $this->createZip($file->children);
                     $filename = $file->name . '.zip';
                 } else {
-                    $path = 'public/' . pathinfo($file->storage_path, PATHINFO_BASENAME);
-                    $dest = $path;
+                    $dest = 'public/' . pathinfo($file->storage_path, PATHINFO_BASENAME);
                     Storage::copy($file->storage_path, $dest);
 
-                    $url = asset(Storage::url($dest, $file->name));
+                    $url = asset(Storage::url($dest));
                     $filename = $file->name;
                 }
             } else {
@@ -200,55 +180,42 @@ class FileController extends Controller
     /**
      *
      *
-     * @return \Illuminate\Database\Eloquent\Builder|\Illuminate\Database\Eloquent\Model
+     * @param $file
+     * @param $user
+     * @param $parent
      * @author Zura Sekhniashvili <zurasekhniashvili@gmail.com>
      */
-    private function getRoot(): \Illuminate\Database\Eloquent\Builder|\Illuminate\Database\Eloquent\Model
+    private function saveFile($file, $user, $parent): void
     {
-        return File::query()->whereIsRoot()->where('created_by', '=', request()->user()->id)->firstOrFail();
+        $path = $file->store('/files/' . $user->id);
+
+        $model = new File();
+        $model->storage_path = $path;
+        $model->is_folder = false;
+        $model->name = $file->getClientOriginalName();
+        $model->mime = $file->getMimeType();
+        $model->size = $file->getSize();
+
+        $parent->appendNode($model);
     }
 
-    private function saveFileTree($fileTree, $parent, $user)
+    public function createZip($files): string
     {
-        foreach ($fileTree as $name => $file) {
-            if (is_array($file)) {
-                $folder = new File();
-                $folder->is_folder = true;
-                $folder->name = $name;
-
-                $parent->appendNode($folder);
-                $this->saveFileTree($file, $folder, $user);
-            } else {
-                $path = $file->store('/files/' . $user->id);
-                $model = new File();
-                $model->storage_path = $path;
-                $model->is_folder = false;
-                $model->name = $file->getClientOriginalName();
-                $model->mime = $file->getMimeType();
-                $model->size = $file->getSize();
-
-                $parent->appendNode($model);
-            }
-        }
-    }
-
-    private function createZip($files): string
-    {
-        // Path to the output ZIP file
         $zipPath = 'zip/' . Str::random() . '.zip';
         $publicPath = "public/$zipPath";
 
         if (!is_dir(dirname($publicPath))) {
             Storage::makeDirectory(dirname($publicPath));
         }
+
         $zipFile = Storage::path($publicPath);
 
-        $zip = new ZipArchive();
-        if ($zip->open($zipFile, ZipArchive::CREATE | ZipArchive::OVERWRITE) === true) {
+        $zip = new \ZipArchive();
+
+        if ($zip->open($zipFile, \ZipArchive::CREATE | \ZipArchive::OVERWRITE) === true) {
             $this->addFilesToZip($zip, $files);
         }
 
-        // Close the ZIP archive
         $zip->close();
 
         return asset(Storage::url($zipPath));
